refactor(spring)!: hoist shared starter wiring into spring-support
The JUnit, Spock, and Kotest Spring Boot starters each reimplemented an
identical DbTesterProperties (~700 lines), the properties-to-Configuration
mapping, and a DataSourceRegistrar. Consolidate them into db-tester-spring-support:

- Move DbTesterProperties (the canonical @ConfigurationProperties class),
  DataSourceRegistrar, and the additional configuration metadata into
  spring-support, and add DbTesterConfigurationFactory for the mapping.
- Reduce each starter to a thin auto-configuration that delegates to the shared
  factory and registrar.
- Remove the dead dbTesterDataSourceRegistry bean and its redundant eager
  registerDefault; each extension now builds a fresh registry and lets the
  registrar resolve the default (single, @Primary, then "dataSource").
- Cache the Spock TestContextManager per specification instead of recreating it
  for every feature method.
- Narrow the DataSourceRegistrarSupport helper methods to package-private.

The db-tester-spring-support module now depends on spring-boot-autoconfigure to
host the @ConfigurationProperties class. The db-tester relaxed binding prefix
and property names are unchanged; the starter context tests are the safety net.

BREAKING CHANGE: DbTesterProperties and DataSourceRegistrar move from each
starter's autoconfigure package to io.github.seijikohara.dbtester.spring.support.
The per-starter dbTesterDataSourceRegistry bean is removed.
