+++ db-tester-junit-spring-boot-starter/src/main/java/io/github/seijikohara/dbtester/junit/spring/boot/autoconfigure/DbTesterJUnitAutoConfiguration.java
@@ -1,18 +1,12 @@
 package io.github.seijikohara.dbtester.junit.spring.boot.autoconfigure;
 
-import io.github.seijikohara.dbtester.api.config.ColumnStrategyMapping;
 import io.github.seijikohara.dbtester.api.config.Configuration;
-import io.github.seijikohara.dbtester.api.config.ConventionSettings;
 import io.github.seijikohara.dbtester.api.config.DataSourceRegistry;
-import io.github.seijikohara.dbtester.api.config.ExecutionSettings;
-import io.github.seijikohara.dbtester.api.config.OperationDefaults;
-import io.github.seijikohara.dbtester.api.config.VerificationSettings;
 import io.github.seijikohara.dbtester.junit.jupiter.extension.DatabaseTestExtension;
-import io.github.seijikohara.dbtester.spring.support.ColumnStrategyConverter;
-import java.util.Map;
-import java.util.stream.Collectors;
+import io.github.seijikohara.dbtester.spring.support.DataSourceRegistrar;
+import io.github.seijikohara.dbtester.spring.support.DbTesterConfigurationFactory;
+import io.github.seijikohara.dbtester.spring.support.DbTesterProperties;
 import javax.sql.DataSource;
-import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
@@ -70,88 +64,7 @@ public class DbTesterJUnitAutoConfiguration {
   @Bean
   @ConditionalOnMissingBean
   public Configuration dbTesterConfiguration(final DbTesterProperties properties) {
-    final DbTesterProperties.ConventionProperties conventionProps = properties.getConvention();
-    final DbTesterProperties.VerificationProperties verificationProps =
-        properties.getVerification();
-    final DbTesterProperties.ExecutionProperties executionProps = properties.getExecution();
-    final DbTesterProperties.OperationProperties operationProps = properties.getOperation();
-
-    final ConventionSettings conventions =
-        ConventionSettings.builder()
-            .baseDirectory(conventionProps.getBaseDirectory())
-            .expectationSuffix(conventionProps.getExpectationSuffix())
-            .scenarioMarker(conventionProps.getScenarioMarker())
-            .dataFormat(conventionProps.getDataFormat())
-            .tableMergeStrategy(conventionProps.getTableMergeStrategy())
-            .loadOrderFileName(conventionProps.getLoadOrderFileName())
-            .build();
-
-    final Map<String, ColumnStrategyMapping> columnStrategies =
-        verificationProps.getColumnStrategies().stream()
-            .filter(
-                prop ->
-                    prop.getColumnName() != null
-                        && !prop.getColumnName().isBlank()
-                        && prop.getStrategy() != null)
-            .map(
-                prop -> {
-                  final var mapping =
-                      ColumnStrategyConverter.toColumnStrategyMapping(
-                          prop.getColumnName(), prop.getStrategy(), prop.getPattern());
-                  return ColumnStrategyConverter.toMapEntry(prop.getColumnName(), mapping);
-                })
-            .collect(
-                Collectors.toMap(
-                    Map.Entry::getKey, Map.Entry::getValue, (first, second) -> second));
-
-    final VerificationSettings verification =
-        VerificationSettings.builder()
-            .globalExcludeColumns(verificationProps.getGlobalExcludeColumns())
-            .globalColumnStrategies(columnStrategies)
-            .rowOrdering(verificationProps.getRowOrdering())
-            .retryCount(verificationProps.getRetryCount())
-            .retryDelay(verificationProps.getRetryDelay())
-            .build();
-
-    final ExecutionSettings execution =
-        ExecutionSettings.builder()
-            .queryTimeout(executionProps.getQueryTimeout())
-            .transactionMode(executionProps.getTransactionMode())
-            .build();
-
-    final OperationDefaults operations =
-        OperationDefaults.builder()
-            .preparation(operationProps.getPreparation())
-            .expectation(operationProps.getExpectation())
-            .build();
-
-    return Configuration.builder()
-        .conventions(conventions)
-        .verification(verification)
-        .execution(execution)
-        .operations(operations)
-        .build();
-  }
-
-  /**
-   * Creates a DataSourceRegistry bean and registers all available DataSources.
-   *
-   * <p>If a single DataSource is available, it will be registered as the default. If multiple
-   * DataSources are available, they will be registered by their bean names.
-   *
-   * @param dataSources provider for all DataSource beans
-   * @return the data source registry
-   */
-  @Bean
-  @ConditionalOnMissingBean
-  public DataSourceRegistry dbTesterDataSourceRegistry(
-      final ObjectProvider<DataSource> dataSources) {
-    final DataSourceRegistry registry = new DataSourceRegistry();
-
-    // Register the first DataSource as default
-    dataSources.stream().findFirst().ifPresent(registry::registerDefault);
-
-    return registry;
+    return DbTesterConfigurationFactory.toConfiguration(properties);
   }
 
   /**
