--- db-tester-kotest-spring-boot-starter/src/main/kotlin/io/github/seijikohara/dbtester/kotest/spring/boot/autoconfigure/DbTesterProperties.kt
@@ -1,146 +1,0 @@
-package io.github.seijikohara.dbtester.kotest.spring.boot.autoconfigure
-
-import io.github.seijikohara.dbtester.api.config.ConventionSettings
-import io.github.seijikohara.dbtester.api.config.DataFormat
-import io.github.seijikohara.dbtester.api.config.RowOrdering
-import io.github.seijikohara.dbtester.api.config.TableMergeStrategy
-import io.github.seijikohara.dbtester.api.config.TransactionMode
-import io.github.seijikohara.dbtester.api.domain.Strategy
-import io.github.seijikohara.dbtester.api.operation.Operation
-import org.springframework.boot.context.properties.ConfigurationProperties
-import org.springframework.boot.context.properties.NestedConfigurationProperty
-import java.time.Duration
-
-/**
- * Configuration properties for DB Tester Spring Boot integration.
- *
- * These properties control how DataSource beans from the Spring application context are
- * registered with the [io.github.seijikohara.dbtester.api.config.DataSourceRegistry], as well
- * as default conventions and operations for database testing.
- *
- * Properties are prefixed with `db-tester`.
- *
- * ## Available Properties
- *
- * - `db-tester.enabled` - Enable/disable DB Tester (default: true)
- * - `db-tester.auto-register-data-sources` - Auto-register DataSources (default: true)
- * - `db-tester.convention.*` - Convention settings for dataset resolution
- * - `db-tester.operation.*` - Default database operations
- */
-@ConfigurationProperties(prefix = "db-tester")
-class DbTesterProperties {
-    /** Whether DB Tester is enabled. Defaults to true. */
-    var enabled: Boolean = true
-
-    /** Whether to automatically register DataSource beans. Defaults to true. */
-    var autoRegisterDataSources: Boolean = true
-
-    /** Convention settings for dataset resolution. */
-    @NestedConfigurationProperty
-    var convention: ConventionProperties = ConventionProperties()
-
-    /** Verification settings for expectation behavior. */
-    @NestedConfigurationProperty
-    var verification: VerificationProperties = VerificationProperties()
-
-    /** Execution settings for database operation behavior. */
-    @NestedConfigurationProperty
-    var execution: ExecutionProperties = ExecutionProperties()
-
-    /** Default operation settings for preparation and expectation phases. */
-    @NestedConfigurationProperty
-    var operation: OperationProperties = OperationProperties()
-
-    /**
-     * Convention properties for dataset resolution.
-     *
-     * These properties define how dataset files are located and processed.
-     */
-    class ConventionProperties {
-        /** Base directory for dataset resolution. Null means resolve from classpath relative to test class. */
-        var baseDirectory: String? = null
-
-        /** Suffix appended to preparation path for expectation datasets. */
-        var expectationSuffix: String = ConventionSettings.DEFAULT_EXPECTATION_SUFFIX
-
-        /** Column name that identifies scenario markers in dataset files. */
-        var scenarioMarker: String = ConventionSettings.DEFAULT_SCENARIO_MARKER
-
-        /** File format for dataset files. Defaults to AUTO. */
-        var dataFormat: DataFormat = DataFormat.AUTO
-
-        /** Strategy for merging tables from multiple datasets. Defaults to UNION_ALL. */
-        var tableMergeStrategy: TableMergeStrategy = TableMergeStrategy.UNION_ALL
-
-        /** File name for specifying table loading order in dataset directories. */
-        var loadOrderFileName: String = ConventionSettings.DEFAULT_LOAD_ORDER_FILE_NAME
-    }
-
-    /**
-     * Verification properties for expectation behavior.
-     *
-     * These properties control how the framework verifies expected database state after test
-     * execution.
-     */
-    class VerificationProperties {
-        /** Column names to exclude globally from all expectation verifications. */
-        var globalExcludeColumns: Set<String> = emptySet()
-
-        /** Column-level comparison strategies for expectation verification. */
-        var columnStrategies: MutableList<ColumnStrategyProperty> = mutableListOf()
-
-        /** Row ordering strategy for expectation verification. Defaults to ORDERED. */
-        var rowOrdering: RowOrdering = RowOrdering.ORDERED
-
-        /** Number of retry attempts for expectation verification. Defaults to 0. */
-        var retryCount: Int = 0
-
-        /** Delay between retry attempts. Defaults to 100ms. */
-        var retryDelay: Duration = Duration.ofMillis(100)
-    }
-
-    /**
-     * Property for configuring a column-level comparison strategy.
-     *
-     * Each entry associates a column name with a [Strategy] and an optional
-     * pattern for regex-based strategies.
-     */
-    class ColumnStrategyProperty {
-        /** The column name for the strategy. */
-        var columnName: String? = null
-
-        /** The comparison strategy type. */
-        var strategy: Strategy? = null
-
-        /** The regex pattern for REGEX strategy. */
-        var pattern: String? = null
-    }
-
-    /**
-     * Execution properties for database operation behavior.
-     *
-     * These properties control how the framework executes database operations during test
-     * preparation.
-     */
-    class ExecutionProperties {
-        /** Maximum time to wait for database queries. Null means no timeout. */
-        var queryTimeout: Duration? = null
-
-        /** Transaction behavior for database operations. Defaults to SINGLE_TRANSACTION. */
-        var transactionMode: TransactionMode = TransactionMode.SINGLE_TRANSACTION
-    }
-
-    /**
-     * Operation properties for database test phases.
-     *
-     * These properties define the default operations executed during test preparation and
-     * expectation verification.
-     */
-    class OperationProperties {
-        /** Default operation for test preparation phase. Defaults to CLEAN_INSERT. */
-        var preparation: Operation = Operation.CLEAN_INSERT
-
-        /** Default operation for expectation verification phase. Defaults to NONE. */
-        var expectation: Operation = Operation.NONE
-    }
-}
