--- db-tester-spock-spring-boot-starter/src/main/groovy/io/github/seijikohara/dbtester/spock/spring/boot/autoconfigure/DbTesterProperties.groovy
@@ -1,208 +1,0 @@
-package io.github.seijikohara.dbtester.spock.spring.boot.autoconfigure
-
-import io.github.seijikohara.dbtester.api.config.ConventionSettings
-import io.github.seijikohara.dbtester.api.config.DataFormat
-import io.github.seijikohara.dbtester.api.config.RowOrdering
-import io.github.seijikohara.dbtester.api.config.TableMergeStrategy
-import io.github.seijikohara.dbtester.api.config.TransactionMode
-import io.github.seijikohara.dbtester.api.domain.Strategy
-import io.github.seijikohara.dbtester.api.operation.Operation
-import java.time.Duration
-import org.springframework.boot.context.properties.ConfigurationProperties
-import org.springframework.boot.context.properties.NestedConfigurationProperty
-
-/**
- * Configuration properties for DB Tester Spring Boot integration with Spock.
- *
- * <p>These properties control how DataSource beans from the Spring application context are
- * registered with the {@link io.github.seijikohara.dbtester.api.config.DataSourceRegistry}, as well
- * as default conventions and operations for database testing.
- *
- * <p>Properties are prefixed with {@code db-tester}.
- *
- * <h2>Available Properties</h2>
- *
- * <ul>
- *   <li>{@code db-tester.enabled} - Enable/disable DB Tester (default: true)
- *   <li>{@code db-tester.auto-register-data-sources} - Auto-register DataSources (default: true)
- *   <li>{@code db-tester.convention.*} - Convention settings for dataset resolution
- *   <li>{@code db-tester.operation.*} - Default database operations
- * </ul>
- */
-@ConfigurationProperties(prefix = "db-tester")
-class DbTesterProperties {
-
-	/** Whether DB Tester is enabled. Defaults to true. */
-	boolean enabled = true
-
-	/** Whether to automatically register DataSource beans. Defaults to true. */
-	boolean autoRegisterDataSources = true
-
-	/** Convention settings for dataset resolution. */
-	@NestedConfigurationProperty
-	ConventionProperties convention = new ConventionProperties()
-
-	/** Verification settings for expectation behavior. */
-	@NestedConfigurationProperty
-	VerificationProperties verification = new VerificationProperties()
-
-	/** Execution settings for database operation behavior. */
-	@NestedConfigurationProperty
-	ExecutionProperties execution = new ExecutionProperties()
-
-	/** Default operation settings for preparation and expectation phases. */
-	@NestedConfigurationProperty
-	OperationProperties operation = new OperationProperties()
-
-	/**
-	 * Convention properties for dataset resolution.
-	 *
-	 * <p>These properties define how dataset files are located and processed.
-	 *
-	 * <h2>Available Properties</h2>
-	 *
-	 * <ul>
-	 *   <li>{@code db-tester.convention.base-directory} - Base directory for datasets (default: null,
-	 *       uses classpath)
-	 *   <li>{@code db-tester.convention.expectation-suffix} - Suffix for expectation directories
-	 *       (default: /expected)
-	 *   <li>{@code db-tester.convention.scenario-marker} - Column name for scenario filtering
-	 *       (default: [Scenario])
-	 *   <li>{@code db-tester.convention.data-format} - Dataset file format (default: AUTO)
-	 *   <li>{@code db-tester.convention.table-merge-strategy} - Strategy for merging tables (default:
-	 *       UNION_ALL)
-	 *   <li>{@code db-tester.convention.load-order-file-name} - File name for table loading order
-	 *       (default: load-order.txt)
-	 * </ul>
-	 */
-	static class ConventionProperties {
-
-		/**
-		 * Base directory for dataset resolution. Null means resolve from classpath relative to test
-		 * class.
-		 */
-		String baseDirectory
-
-		/** Suffix appended to preparation path for expectation datasets. Defaults to "/expected". */
-		String expectationSuffix = ConventionSettings.DEFAULT_EXPECTATION_SUFFIX
-
-		/** Column name that identifies scenario markers in dataset files. Defaults to "[Scenario]". */
-		String scenarioMarker = ConventionSettings.DEFAULT_SCENARIO_MARKER
-
-		/** File format for dataset files. Defaults to AUTO. */
-		DataFormat dataFormat = DataFormat.AUTO
-
-		/** Strategy for merging tables from multiple datasets. Defaults to UNION_ALL. */
-		TableMergeStrategy tableMergeStrategy = TableMergeStrategy.UNION_ALL
-
-		/** File name for specifying table loading order in dataset directories. Defaults to "load-order.txt". */
-		String loadOrderFileName = ConventionSettings.DEFAULT_LOAD_ORDER_FILE_NAME
-	}
-
-	/**
-	 * Verification properties for expectation behavior.
-	 *
-	 * <p>These properties control how the framework verifies expected database state after test
-	 * execution.
-	 *
-	 * <h2>Available Properties</h2>
-	 *
-	 * <ul>
-	 *   <li>{@code db-tester.verification.global-exclude-columns} - Column names to exclude globally
-	 *       (default: empty)
-	 *   <li>{@code db-tester.verification.column-strategies} - Column-level comparison strategies
-	 *       (default: empty)
-	 *   <li>{@code db-tester.verification.row-ordering} - Row ordering strategy (default: ORDERED)
-	 *   <li>{@code db-tester.verification.retry-count} - Retry attempts (default: 0)
-	 *   <li>{@code db-tester.verification.retry-delay} - Delay between retries (default: 100ms)
-	 * </ul>
-	 */
-	static class VerificationProperties {
-
-		/** Column names to exclude globally from all expectation verifications. */
-		Set<String> globalExcludeColumns = Set.of()
-
-		/** Column-level comparison strategies for expectation verification. */
-		List<ColumnStrategyProperty> columnStrategies = []
-
-		/** Row ordering strategy for expectation verification. Defaults to ORDERED. */
-		RowOrdering rowOrdering = RowOrdering.ORDERED
-
-		/** Number of retry attempts for expectation verification. Defaults to 0. */
-		int retryCount = 0
-
-		/** Delay between retry attempts. Defaults to 100ms. */
-		Duration retryDelay = Duration.ofMillis(100)
-	}
-
-	/**
-	 * Property for configuring a column-level comparison strategy.
-	 *
-	 * <p>Each entry associates a column name with a {@link Strategy} and an optional
-	 * pattern for regex-based strategies.
-	 *
-	 * <p>Example usage in {@code application.properties}:
-	 *
-	 * <pre>{@code
-	 * db-tester.verification.column-strategies[0].column-name=CREATED_AT
-	 * db-tester.verification.column-strategies[0].strategy=TIMESTAMP_FLEXIBLE
-	 * }</pre>
-	 */
-	static class ColumnStrategyProperty {
-
-		/** The column name for the strategy. */
-		String columnName
-
-		/** The comparison strategy type. */
-		Strategy strategy
-
-		/** The regex pattern for REGEX strategy. */
-		String pattern
-	}
-
-	/**
-	 * Execution properties for database operation behavior.
-	 *
-	 * <p>These properties control how the framework executes database operations during test
-	 * preparation.
-	 *
-	 * <h2>Available Properties</h2>
-	 *
-	 * <ul>
-	 *   <li>{@code db-tester.execution.query-timeout} - Maximum query wait time (default: none)
-	 *   <li>{@code db-tester.execution.transaction-mode} - Transaction behavior (default:
-	 *       SINGLE_TRANSACTION)
-	 * </ul>
-	 */
-	static class ExecutionProperties {
-
-		/** Maximum time to wait for database queries. Null means no timeout. */
-		Duration queryTimeout = null
-
-		/** Transaction behavior for database operations. Defaults to SINGLE_TRANSACTION. */
-		TransactionMode transactionMode = TransactionMode.SINGLE_TRANSACTION
-	}
-
-	/**
-	 * Operation properties for database test phases.
-	 *
-	 * <p>These properties define the default operations executed during test preparation and
-	 * expectation verification.
-	 *
-	 * <h2>Available Properties</h2>
-	 *
-	 * <ul>
-	 *   <li>{@code db-tester.operation.preparation} - Default preparation operation (default:
-	 *       CLEAN_INSERT)
-	 *   <li>{@code db-tester.operation.expectation} - Default expectation operation (default: NONE)
-	 * </ul>
-	 */
-	static class OperationProperties {
-
-		/** Default operation for test preparation phase. Defaults to CLEAN_INSERT. */
-		Operation preparation = Operation.CLEAN_INSERT
-
-		/** Default operation for expectation verification phase. Defaults to NONE. */
-		Operation expectation = Operation.NONE
-	}
-}
