+++ db-tester-spring-support/src/main/java/io/github/seijikohara/dbtester/spring/support/DbTesterProperties.java
@@ -1,4 +1,4 @@
-package io.github.seijikohara.dbtester.junit.spring.boot.autoconfigure;
+package io.github.seijikohara.dbtester.spring.support;
 
 import io.github.seijikohara.dbtester.api.config.ConventionSettings;
 import io.github.seijikohara.dbtester.api.config.DataFormat;
