+++ db-tester-spring-support/src/test/java/io/github/seijikohara/dbtester/spring/support/DataSourceRegistrarTest.java
@@ -1,4 +1,4 @@
-package io.github.seijikohara.dbtester.junit.spring.boot.autoconfigure;
+package io.github.seijikohara.dbtester.spring.support;
 
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertFalse;
